build(dsl): Add the kotlin.internal.OnlyInputTypes annotation
It is required to avoid upcasting by the compiler in cases like:
```kotlin
Foo::anInt eq "foo"
```
By default, the Kotlin compiler will upcast both values, so it will consider `eq` to be called with the type parameter `Any`.

That is barely useful to us, though. Using `OnlyInputTypes` we can refuse upcasting. Instead, the compiler will refuse the above code, which is what we want.