refactor(dsl): Migrate to the multi-dollar interpolation syntax
Kotlin uses the dollar sign for interpolation syntax. In multi-line strings, the dollar sign cannot be escaped. In KtMongo, we use a ton of multi-line strings, and BSON uses the dollar sign to mark operators.

Previously, we used a basic workaround by declaring a variable which did contain the literal:
```kotlin
val foo = "$foo"
println("""
    { "$foo": 2 }
""".trimIndent())
```
This workaround meant that requests written in the code would look like the BSON syntax even if that technically wasn't possible in Kotlin. However, this had the downside of needing to import each such variable, and being inconvenient to write.

In Kotlin 2.2.0, we can use two dollar signs (or more) in front a string literal to specify that interpolation needs that many dollar signs. Fewer dollar signs are just the literal dollar symbol.

We can thus write:
```kotlin
println($$"""
    { "$foo": 2 }
""".trimIndent())
```
Because '$foo' is a single dollar sign, and the string is prefixed with two dollar signs, then it's not interpolation.

See https://github.com/Kotlin/KEEP/issues/375